fix: error deleting customers without invoices

+++ frontend/app/shared/infrastructure/db/InvoiceRepositoryImpl.ts
@@ -8,7 +8,9 @@ export class InvoiceRepositoryImpl implements InvoiceRepository {
 
     async deleteInvoiceByCustomerId(customerId: number): Promise<void> {
         const invoices = await this.db.selectFrom('invoice').where('customerId', '=', customerId).select('invoiceId').execute()
-        await this.db.deleteFrom('invoiceLine').where('invoiceId', 'in', [...invoices.flatMap((v) => v.invoiceId)]).execute()
-        await this.db.deleteFrom('invoice').where('customerId', '=', customerId).execute()
+        if (invoices?.length > 0) {
+            await this.db.deleteFrom('invoiceLine').where('invoiceId', 'in', [...invoices.flatMap((v) => v.invoiceId)]).execute()
+            await this.db.deleteFrom('invoice').where('customerId', '=', customerId).execute()
+        }
     }
 }
